Treat Resource Group as a String not an Object
A previous change assumed that a PR had been backported from Fine to upstream converting
Resource Groups to objects which had to be dereferenced to get to the name attribute.
This PR had not been backported so we need to treat the value as a String.

This is for high priority BZ https://bugzilla.redhat.com/show_bug.cgi?id=1475540
and goes hand in hand with https://github.com/ManageIQ/manageiq-gems-pending/pull/273

This is for FINE only.