Cache availability check for 1 hour
This check ends up getting called once every 15 seconds by the
sync_workers loop, which is expensive considering it rarely changes.
This cache is a temporary fix to lighten the CPU load on the
orchestrator, where it runs, until such time as we can properly cache
the value in the database.

Partially fixes #438