Improve Invalid Credentials error message
There are four components that make up Azure credentials: tenant_id,
subscription_id, client_id, and client_secret.

If you have the wrong subscription ID there is a different exception
that is raised:
`#<ArgumentError: Subscription ID '2b01dec2-6684-4f25-abfc-a2453e0ae52e' not found>`

But if the Tenant ID, Client ID, or Client Secret are wrong we just get:
`#<Azure::Armrest::BadRequestException: [] Azure::Armrest::BadRequestException (cause: 400 Bad Request)>`

We were not indicating in the error message that the tenant ID could be
what is wrong.

We are also not able to indicate _which_ of the three is wrong based on
the error.