Fix `can't cast hash` error during provisioning
When provisioning an instance a targeted refresh is queued using the
value returned by the `start_clone` method as the `:ems_ref`.

Azure had been returning a hash rather than a string which was causing
the Refresher to fail when querying for existing records using this hash
as the `:ems_ref`.

The solution is to simply build the ems_ref the same way the Azure
refresher does by joining the subscription, resource_group, type, and vm
name.